Inline tests.
